+++ anywise/_registry.py
@@ -1,10 +1,12 @@
 import inspect
-from typing import Any, Callable, overload
+from collections import defaultdict
+from typing import Callable, overload
 
 from ididi import DependencyGraph, INode
 
 from ._itypes import FuncMeta, HandlerMapping, ListenerMapping, MethodMeta
 from ._visitor import Target, collect_handlers, collect_listeners
+from .guard import Guard, GuardFunc, PostHandle
 
 # def auto_collect(msg_type: type, dir: pathlib.Path):
 #     """
@@ -14,6 +16,17 @@ from ._visitor import Target, collect_handlers, collect_listeners
 #     constrcut a anywise.pyi stub file along the way
 #     """
 #     ...
+# from functools import lru_cache
+
+
+# @lru_cache
+def handler_registry[C](msg_type: type[C]) -> "HandlerRegistry[C]":
+    return HandlerRegistry[C](msg_type)
+
+
+# @lru_cache
+def listener_registry[E](msg_type: type[E]) -> "ListenerRegistry[E]":
+    return ListenerRegistry[E](msg_type)
 
 
 class RegistryBase[Message]:
@@ -21,9 +34,7 @@ class RegistryBase[Message]:
         self, message_type: type[Message], *, graph: DependencyGraph | None = None
     ):
         self._message_type = message_type
-        if graph is None:
-            graph = DependencyGraph()
-        self._graph = graph
+        self._graph = graph or DependencyGraph()
 
     @property
     def graph(self) -> DependencyGraph:
@@ -32,9 +43,6 @@ class RegistryBase[Message]:
     def factory[**P, R](self, factory: INode[P, R]) -> INode[P, R]:
         self._graph.node(factory)
         return factory
-
-    def guard(self, func: Any):
-        "like middleware in starlette"
 
     @overload
     def register[R](self, handler: type[R]) -> type[R]: ...
@@ -90,11 +98,18 @@ class ListenerRegistry[Event](RegistryBase[Event]):
                     self._graph.node(ignore=msg_type)(meta.owner_type)
                 else:
                     listener = meta.handler
-                    entry = self._graph.entry(ignore=msg_type)(listener)
+
+                    if meta.is_contexted:
+                        ignore = (msg_type, "context")
+                    else:
+                        ignore = msg_type
+
+                    entry = self._graph.entry(ignore=ignore)(listener)
                     metas[i] = FuncMeta(
                         message_type=msg_type,
                         handler=entry,
                         is_async=inspect.iscoroutinefunction(listener),
+                        is_contexted=meta.is_contexted,
                     )
             self._mapping[msg_type].extend(metas)
 
@@ -108,8 +123,8 @@ class HandlerRegistry[Command](RegistryBase[Command]):
 
     def __iter__(self):
         items = self._mapping.items()
-        for msg_type, handler_meta in items:
-            yield (msg_type, handler_meta)
+        for msg_type, funcmeta in items:
+            yield (msg_type, funcmeta)
 
     @overload
     def register[T](self, handler: type[T]) -> type[T]: ...
@@ -125,23 +140,92 @@ class HandlerRegistry[Command](RegistryBase[Command]):
                 self._graph.node(ignore=msg_type)(meta.owner_type)
             else:
                 entry = self._graph.entry(ignore=msg_type)(f)
+
                 mappings[msg_type] = FuncMeta(
                     message_type=msg_type,
                     handler=entry,
                     is_async=inspect.iscoroutinefunction(f),
+                    is_contexted=meta.is_contexted,
                 )
         self._mapping.update(mappings)
         return handler
 
 
-# from functools import lru_cache
+# class GuardRegistry[Command](RegistryBase[Command]):
+#     def __init__(self, message_type: type[Command]):
+#         super().__init__(message_type=message_type)
+#         self._mapping: defaultdict[Command, list[GuardFunc]] = defaultdict(list)
+
+#     def guard(self, message_type: type):
+#         "like middleware in starlette"
+
+#         """
+#         register guard into guard registry
+#         when included in anywise, match handler by command type
+#         a guard of base command will be added to all handlers of subcommand, meaning
+
+#         guard(UserCommand)
+
+#         will be added to handle of CreateUser, UpdateUser, etc.
 
 
-# @lru_cache
-def handler_registry[C](msg_type: type[C]) -> HandlerRegistry[C]:
-    return HandlerRegistry[C](msg_type)
+#         class AuthService:
+#             @guard(UserCommand)
+#             def validate_user(self, command: UserCommand, context: AuthContext):
+#                 user = self._get_user(context.token.sub)
+#                 if user.user_id != command.user_id:
+#                     raise ValidationError
+#                 context["user"] = user
+
+#         @guard.pre_handle
+#         def validate_user(self): ...
 
 
-# @lru_cache
-def listener_registry[E](msg_type: type[E]) -> ListenerRegistry[E]:
-    return ListenerRegistry[E](msg_type)
+#         auth_guard = guard(AuthGuard, *arsg, **kwargs)
+#         logging_guard = guard(LoggingGuard, *args, **kwargs)
+
+#         auth_guard.bind([UserCommand, ProductCommand])
+#         """
+
+#     def pre_handler(self, func: GuardFunc):
+#         self._mapping[self._message_type] = Guard(pre_handle=func)
+
+#     def bind(self, message_type: type | list[type]):
+#         """
+#         which command this guard belongs to
+#         """
+
+
+class GuardRegistry:
+    def __init__(self):
+        # self._pre_handlers: defaultdict[type, list[GuardFunc]] = defaultdict(list)
+        # self._post_hanlders: defaultdict[type, list[PostHandle]] = defaultdict(list)
+        self._guards: defaultdict[type, list[Guard]] = defaultdict(list)
+        self.graph = None
+
+    def __iter__(self):
+        mappings = self._guards.items()
+
+        for msg_type, guards in mappings:
+            yield (msg_type, guards)
+
+    # def pre_handle(self, command_type: type):
+    #     def receive(func: GuardFunc) -> GuardFunc:
+    #         self._pre_handlers[command_type].append(func)
+    #         return func
+
+    #     return receive
+
+    # def post_handle(self, command_type: type):
+    #     def receive(func: PostHandle) -> PostHandle:
+    #         self._post_hanlders[command_type].append(func)
+    #         return func
+
+    #     return receive
+
+    def register(self, command_type: type):
+        def receive(func: GuardFunc):
+            self._guards[command_type].append(Guard(pre_handle=func))
+            return func
+
+        return receive