+++ anywise/publisher.py
@@ -8,4 +8,4 @@ class ConcurrentPublisher(Publisher):
     async def publish(self, msg: Any) -> None:
         subscribers = self._subscribers[type(msg)]
         async with TaskGroup() as tg:
-            [tg.create_task(sub.handle(msg)) for sub in subscribers]
+            [tg.create_task(sub(msg)) for sub in subscribers]