+++ tests/test_guard.py
@@ -1,46 +1,30 @@
 import typing as ty
 
-from anywise.guard import MarkGuard
+from anywise import AnyWise, handler_registry
+
+# from anywise.guard import MarkGuard
+from anywise._registry import GuardFunc, GuardRegistry
+from tests.conftest import CreateUser
+
+guard_maker = GuardRegistry()
+cmd_handler = handler_registry(CreateUser)
 
 
-async def handler_str(name: str, context: dict[str, ty.Any]):
-    print(context, "done")
+@cmd_handler
+async def handler_str(create_user: CreateUser, context: dict[str, ty.Any]):
+    assert context["processed_by"]
     return "done"
 
 
-def guard[
-    **P, R
-](message_type: type, guard_class: ty.Callable[P, R] | None = None,):
-    """
-    register guard into guard registry
-    when included in anywise, match handler by command type
-    a guard of base command will be added to all handlers of subcommand, meaning
-
-    guard(UserCommand) will be added to handle of CreateUser, UpdateUser,
-    etc.
-
-
-
-    user_guard = guard(UserCommand)
-
-    @user_guard
-    class AuthService:
-        def validate_user(command: UserCommand, context: AuthContext):
-            user = self._get_user(context.token.sub)
-            if user.user_id != command.user_id:
-                raise ValidationError
-            context["user"] = user
-
-    @user_guard.pre_handle
-    def validate_user(self): ...
-
-    """
-    ...
+@guard_maker.register(CreateUser)  # pre handle
+async def mark(create_user: CreateUser, context: dict[str, ty.Any]) -> None:
+    if not context.get("processed_by"):
+        context["processed_by"] = ["1"]
+    else:
+        context["processed_by"].append("1")
 
 
 async def test_guard():
-    m1 = MarkGuard(MarkGuard(MarkGuard(handler_str)))
-
-    # m1.add_new(m2).add_new(m3)
-    res = await m1("name", dict())
-    assert res == "done"
+    aw = AnyWise()
+    aw.include([cmd_handler, guard_maker])
+    await aw.send(CreateUser("1", "2"))